Added Wifi class (Beta)

+++ app/src/main/java/app_mobili/transceiver_go/Sensor.java
@@ -1,7 +1,7 @@
 package app_mobili.transceiver_go;
 
 public class Sensor {
-    protected float SensorValue;
+    protected int SensorValue;
 
     public Sensor() {
         SensorValue = -1;
